// src/domain/dashboard/balance/types/account-component-modal-types.ts
export interface IAccountComponentModalTypes {
  bank?: string;
  bankName?: string;
  favoredName?: string;
  personType?: string;
  cpf?: string;
  cnpj?: string;
  accountType?: string;
  agency?: string;
  account?: string;
  digit?: string;
  zipCode?: string;
  street?: string;
  number?: string;
  complement?: string;
  district?: string;
  cityName?: string;
  stateInitials?: string;
  countryName?: string;
}